[fixed] adding markers

+++ app/src/main/java/com/example/sander/locationaware/LocationMarker.java
@@ -35,7 +35,7 @@ public class LocationMarker {
         return date;
     }
 
-    public void setDate(int date) {
+    public void setDate(long date) {
         this.date = date;
     }
